% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_risk_quadrant.R
\name{calculate_risk_quadrant}
\alias{calculate_risk_quadrant}
\title{Calculate placement on risk quadrant plot}
\usage{
calculate_risk_quadrant(suit.x, suit.y, thresh.x, thresh.y)
}
\arguments{
\item{suit.x}{Suitability values comprising the x-axis.}

\item{suit.y}{Suitability values comprising the y-axis.}

\item{thresh.x}{The threshold for suitability along the x-axis.}

\item{thresh.y}{The threshold for suitability along the y-axis.}
}
\value{
One of: "extreme", "high", "moderate", or "low". These correspond with our
interpretation of the risk quadrant plots we have created:
\itemize{
\item extreme risk = quadrant 4 (upper right)
\item high risk = quadrant 3 (upper left)
\item moderate risk = quadrant 2 (bottom right)
\item low risk = quadrant 1 (bottom left)
}
}
\description{
This function calculates the risk of establishment by Lycorma
delicatula according to our risk quadrant framework (which quadrant a suitability
value will fall into on an xy-scatter). It requires the input of a matrix of
suitability values and a corresponding suitability threshold for each of the x
and y axes for this plot.
}
\details{
details
}
\examples{
example

}
