% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_DD_suitability.R
\name{compute_DD_suitability}
\alias{compute_DD_suitability}
\title{Compute annual degree day accumulation at listed x,y coordinate locations}
\usage{
compute_DD_suitability(
  x,
  thresh = "adult_emergence_1",
  create.output = FALSE,
  mypath = NA,
  create.dir = FALSE
)
}
\arguments{
\item{x}{Data import. Should be a .csv file or data frame that contains
input temperature data. If a .csv file, file path should be in the format
produced by the \code{\link[=file.path]{file.path()}} function (i.e. with '/' instead of '\\').
See details for additional import formatting.}

\item{thresh}{A numeric value or pre-build character string, which represents
a commonly used threshold (see details for list of pre-build values).
This sets the threshold for the number of Degree Days necessary for a
population of \emph{Lycorma delicatula} to reach the desired life stage.}

\item{create.output}{Logical. If TRUE, will create additional outputs from the
returned dataframe (see "return"). Additional outputs include:
\itemize{
\item a .csv file of the returned dataframe
\item a global-scale .tif raster of the output at the 5 arcminute (~10km) resolution
}}

\item{mypath}{Character. Only used if create.output = TRUE. A file path to the
sub directory where the additional outputs will be stored. Should be
used with the \code{\link[=file.path]{file.path()}} function (i.e. with '/' instead of '\\').
If this sub directory does not already exist and should be created by the
function, set \code{create.dir} = TRUE. This will create a folder from the last
part of the file path in \code{mypath}.}

\item{create.dir}{Logical. Should the last element of \code{mypath} create a sub
directory for the output? If TRUE, the folder name specified will be created
for the output. If FALSE (ie, the sub directory already exists), no directory
will be created.}
}
\value{
A data frame containing the total annual degree days and a binary
measure of that location's suitability for establishment by
\emph{Lycorma delicatula}, per coordinate pair. It will have the same length as
the input and will contain the following columns:
\itemize{
\item line no. = line number
\item cell = grid cell unique identifier
\item x = longitude
\item y = latitude
\item total_annual_dds = total degree days in one calendar year for that
coordinate pair
\item suitability = binary, 1 if total_annual_dds >= thresh,
0 if total_annual_dds < thresh
}

IF \code{create.output = TRUE}

The following additional outputs will be created:
#'* a .csv file of the output dataframe
\itemize{
\item a global-scale .tif raster of the output at the 5 arcminute (~10km) resolution
}
\subsection{Error Messages}{

An error message will be given if the input data \code{x} are in the incorrect
format (see details) or if an incompatible value is given for parameter
\code{thresh} (if value is not a preset character string or a numeric object).
}
}
\description{
The function requires the packages 'tidyverse', 'here', and 'terra'
}
\details{
A function that computes the total annual degree day accumulation for
\emph{Lycorma delicatula} development, based on monthly mean temperature data.
This function was developed from the work of Dennis Calvin on the development
of populations of \emph{Lycorma delicatula} living on \emph{Ailanthus altissima} at
different DD accumulations.
\subsection{x:}{

Data frames / .csv files should contain 15 columns. The initial 3
columns should contain descriptors, with the 1st column as the raster cell
number (see examples) and the 2nd and 3rd columns as longitude and latitude,
respectively. If 1st column is unavailable, see "troubleshooting" section of
\code{examples} for code to produce this. Columns 4-15 should contain mean
temperature data for each month of the year (one month per column), with each
row being a grid cell of the raster that was used to create it. The raster
that was used in the original workflow was at the 5 arcminute scale.
}

\subsection{thresh:}{

We have pre-built a number of important DD thresholds, as outlined by Dennis
Calvin's work. These pre-build values can be called by entering their name
as a character string. Pre-built values for \code{thresh} include the following:
\itemize{
\item "adult_emergence_1" = 991 DD (default value), when 1\% of a population
reaches the adult life stage. \strong{This was outlined as a critical threshold for
establishment in North America.}
\item "adult_emergence_50" = 1115.5 DD, when 50\% of a population reaches the
adult stage.
\item "egg_laying_10" = 1564.1 DD, when 10\% of a population reaches egg-laying.
\item "egg_laying_50" = 1616.4 DD, when 50\% of a population
reaches egg-laying
}

If only using for a data frame output, create as an object, such as:
\code{object <- compute_DD_suitability()}

If multiple thresh outputs are desired, must be in the concatenated form:
\code{c("adult_emergence_1", "egg_laying_10")}. Should be all lowercase.
}
}
\examples{
# x---------------------------------------------------------------------------

#' compute_DD_suitability(x = data.object)


compute_DD_suitability(x = file.path("path", "file.csv"))

# thresh----------------------------------------------------------------------

# DD can be set using a preset value. "egg_laying_10" is a preset value
# equal to 1564.1 DDs

compute_DD_suitability(x = data.object, thresh = "egg_laying_10")


# DD can be set manually to a numeric value

compute_DD_suitability(x = data.object, thresh = 1000)

# troubleshooting-------------------------------------------------------------

# If your input data does not contain a column of cell values (this
# requirement is outlined in the details section), one can easily be created
using the seq() function:

library(tidyverse)

data <- data \%>\%
  mutate(cell = seq(nrow(data.object))) \%>\%
  select(cell, tidyselect::everything())

}
