% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_MaxEnt_fixed_ROC.R
\name{compute_MaxEnt_fixed_ROC}
\alias{compute_MaxEnt_fixed_ROC}
\title{Create a model output}
\usage{
compute_MaxEnt_fixed_ROC(
  model.obj,
  model.name,
  mypath,
  fixed.area.points,
  env.covar.obj,
  test.obj,
  create.dir = FALSE
)
}
\arguments{
\item{model.obj}{A model object created by the package 'SDMtune', should be of
class \code{SDMmodelCV}.}

\item{model.name}{Character. A string matching the name of the object set for
\code{model.obj}. Exclude unnecessary phrases, such as the "_model" ending.}

\item{mypath}{Character.A file path to the sub directory where the model
output will be stored. Should be used with the \code{\link[=file.path]{file.path()}} function
(i.e. with '/' instead of '\\').}

\item{fixed.area.points}{Data import. Should be a .csv file or data frame
that contains the fixed area points. If a .csv file, file path should be in the format
produced by the \code{\link[=file.path]{file.path()}} function (i.e. with '/' instead of '\\').
See details for additional import formatting.}

\item{env.covar.obj}{A stack of rasters of environmental covariates. These
covariates are used to train and test the MaxEnt model, as well as to make
predictions. These should be the same covariates that you used to train the
model. This must a \code{SpatRaster} object created using \code{\link[terra:rast]{terra::rast()}}.}

\item{test.obj}{A withheld group of presence and background points used to
test the model after training. Should be a SWD object, created using the
\code{\link[SDMtune:prepareSWD]{SDMtune::prepareSWD()}} function.}
}
\value{
blarg
}
\description{
DESCRIPTION
}
\details{
The function requires the packages 'tidyverse', 'here', 'devtools', 'viridis', 'pROC', 'SDMtune', and 'patchwork'
\subsection{fixed.area.points:}{

The fixed area points dataset should be be in the following format: 3 columns,
with x as longitude, y as latitude and cell as the cell number from the
raster, in that order. Original dataset was generated using
\code{dismo::randomPoints()}
}
}
\examples{

TBD

}
