% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_MaxEnt_summary_statistics.R
\name{compute_MaxEnt_summary_statistics}
\alias{compute_MaxEnt_summary_statistics}
\title{Create a model output}
\usage{
compute_MaxEnt_summary_statistics(
  model.obj,
  model.name,
  mypath,
  env.covar.obj,
  train.obj,
  trainFolds.obj,
  test.obj,
  plot.types,
  threshold.types
)
}
\arguments{
\item{model.obj}{}

\item{model.name}{}

\item{mypath}{}

\item{env.covar.obj}{}

\item{train.obj}{}

\item{trainFolds.obj}{}

\item{test.obj}{}

\item{plot.types}{}

\item{threshold.types}{}
}
\value{
All MaxEnt Summary Statistics
}
\description{
The function requires the packages 'tidyverse', 'here', 'devtools', 'rJava', 'dismo', 'SDMtune', 'viridis' and 'terra'.
}
\details{
model.obj <- entire_easternUSA_model
model.name <- "entire_easternUSA_model"
mypath <- file.path(here() \%>\%
dirname(),
"maxent/models/slf_easternUSA_entire_step1")
env.covar.obj <- x_env_covariates
train.obj <- entire_easternUSA_train
trainFolds.obj <- entire_easternUSA_trainFolds
test.obj <- entire_easternUSA_test
plot.types <- c("train", "test") # used to produce jackknife plots
threshold.types <- c("cloglog", "logistic") # used to produce marginal and univariate response curves
}
\examples{

TBD

}
