% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_celsius.R
\name{convert_to_celsius}
\alias{convert_to_celsius}
\title{Converts units from tenths of a Kelvin to Celsius}
\usage{
convert_to_celsius(x)
}
\arguments{
\item{x}{Data import.}
}
\value{
A numeric value of temperature in Celsius.
}
\description{
This function was specifically written because some temperature data retrieved
from CHELSA were formatted in Kelvin/10. This was made to be applied to a
raster.
}
\details{
This function takes the input, divides it by 10 and subtracts 273.15.

DISCLAIMER: This function was written specifically to convert the unit of the
monthly averaged CHELSA data. The
\href{https://chelsa-climate.org/wp-admin/download-page/CHELSA_tech_specification_V2.pdf}{technical specifications}
for the CHELSA V2.1 data state that the units are in "K/10". This was
originally taken to mean "Kelvin divided by 10", so the value was multiplied
by 10 and converted to Celsius.  but this did not produce a meaningful value.
After instead dividing by 10 and then converting to Celsius, a meaningful
temperature value in Celsius was obtained. So, this was taken to mean "tenths
of a Kelvin, which would require division.
}
\examples{
# x = numeric value-----------------------------------------------------------

# original temperature to be converted is 1500 "tenths of a Kelvin".
x <- 1500
slfSpread::convert_to_celsius(x)

x = numeric value in data frame-----------------------------------------------



x = raster cell---------------------------------------------------------------

raster1 <- terra::rast(x = file)

# use terra:app as an apply function
terra::app(x = raster1, fun = slfSpread::convert_to_celsius())

}
