% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_MaxEnt_suitability_maps.R
\name{create_MaxEnt_suitability_maps}
\alias{create_MaxEnt_suitability_maps}
\title{DESCRIPTION}
\usage{
create_MaxEnt_suitability_maps(
  model.obj,
  model.name,
  mypath,
  create.dir = FALSE,
  env.covar.obj,
  predict.fun = "mean",
  map.thresh = FALSE,
  thresh = NA,
  summary.file = NA,
  map.style = NA
)
}
\arguments{
\item{model.obj}{A model object created by the package 'SDMtune', should be of
class 'SDMmodelCV'.}

\item{model.name}{Character. A string matching the name of the object set for
\code{model.obj}. Exclude unnecessary phrases, such as the "_model" ending.}

\item{mypath}{Character.A file path to the sub directory where the model
output will be stored. Should be used with the \code{\link[=file.path]{file.path()}} function
(i.e. with '/' instead of '\\'). If this sub directory does not already exist
and should be created by the function, set \code{create.dir} = TRUE. This will
create a folder from the last part of the filepath in \code{mypath}.}

\item{create.dir}{Logical. Should the last element of \code{mypath} create a sub
directory for the model output? If TRUE, the main folder will be created for
the model output. If FALSE (ie, the sub directory already exists), only the
"plots" folder within the model output sub directory will be created.}

\item{env.covar.obj}{A stack of rasters of environmental covariates. These
covariates are used to train and test the MaxEnt model, as well as to make
predictions. These should be the same covariates that you used to train the
model. This must a \code{SpatRaster} object created using \code{\link[terra:rast]{terra::rast()}}.}

\item{predict.fun}{Character. The function to be applied to combine the
iterations of the model when predicting a raster output. Can be one of:
"mean", "sd" (standard deviation) or "max". If multiple are desired, must be
in the form: \code{c("mean", "sd", "max")}}

\item{map.thresh}{Logical, TRUE by default. This function determines if a
thresholded suitability map will be created. If not, output will only consist
of suitability maps of the type specified in \code{predict.fun}.}

\item{thresh}{Numeric or Character. Does not need to be defined if
\code{map.thresh = FALSE} This may be imported manually (numeric), or may be
selected from one of the thresholds for the model (character). If a preset,
the specified mean threshold value for all iterations of the model is chosen.
See details for preset options and other usages.}

\item{summary.file}{Data import. Does not need to be defined if
\code{map.thresh = FALSE}. Should be a .csv file or data frame that contains the
summary statistics output created by
\code{\link[=compute_MaxEnt_summary_statistics]{compute_MaxEnt_summary_statistics()}} (filename ending in
"summary_all_iterations.csv"). If an import, file path should be in the format
produced by the \code{\link[=file.path]{file.path()}} function (i.e. with '/' instead of '\\').}

\item{map.style}{List, default is NA. This is used to apply
ggplot aesthetics to the plot outputs. If specified, the given value should be
a list of ggplot aesthetic options. See examples.}
}
\value{
3 outputs
\itemize{
\item mean
}
}
\description{
The function requires the packages 'tidyverse', 'here', 'devtools', 'SDMtune', 'terra' and 'viridis'.
}
\details{
\subsection{thresh:}{

This can be a single numeric or preset character value. It may also be a
concatenated set of numerics or presets, as in  c(0.2, 0.3) or
c("MTSS", "BTO").
}
}
\examples{
# x---------------------------------------------------------------------------


# troubleshooting-------------------------------------------------------------



}
