% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_slf.R
\name{map_slf}
\alias{map_slf}
\title{Create a \link{leaflet} map of SLF occurrences}
\usage{
map_slf(point_lon, point_lat, point_label, extent = 0.01, zoom = 10)
}
\arguments{
\item{point_lon, point_lat}{}

\item{point_label}{}

\item{x}{Data import. Should be a .csv file or data frame that contains
input temperature data. If a .csv file, file path should be in the format
produced by the \code{\link[=file.path]{file.path()}} function (i.e. with '/' instead of '\\').
See details for additional import formatting.}
}
\value{
A logical value (\code{TRUE} or \code{FALSE}) that answers the question
"is any element of the vector not an NA.
}
\description{
This function is designed to map a set of coordinates of interest using the
leaflet package.
}
\details{
using the leaflet package
and in this package, it is used to map new SLF jump locations. It is designed
to show the point with a specified latitude and longitude margin on all sides
of the point (zoom level). It is suggested that this function be used to
append mapping data to a pre-existing data frame or tibble.

#'The function requires the packages 'tidyverse' and 'leaflet'.

This function was previously called map_jumps_v2. Major changes from v1:
Specified all variables as arguments- allows this function to be used in
context of a single point, rather than just a data frame or tibble.
}
\examples{

library(leaflet)
library(tidyverse)

map_slf <-


}
