% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pestHub_compile.R
\name{pestHub_compile}
\alias{pestHub_compile}
\title{Compile data from different files exported by Web Scraping}
\usage{
pestHub_compile(
  filepath,
  columns = "all",
  types,
  counties,
  error.return = FALSE
)
}
\arguments{
\item{filepath}{The file path to the folder that holds the exported CSV files from "scrape_location.ipynb".
Should be in the format produced by the file.path() function (i.e. with '/' instead of '\\')}

\item{columns}{A character or vector of the column names that have the data of interest}

\item{types}{A vector of the location type names (spaces should be used between words)}

\item{counties}{A vector of the county names (spaces should be used between words)}

\item{error.return}{Logical. If \code{error.return = TRUE}, then creates a list of files containing no data as an output}
}
\value{
A data frame of the selected data from the files exported from scrape_location.ipynb.
If \code{error.return = TRUE}, a list is returned with data in the first element and errors in the second element.
}
\description{
This function is used to compile data exported from Web Scraping as CSV files into a single data frame that holds
locations for pest dispersion hubs
}
\details{
The function requires the package 'readr'
}
\examples{
\dontrun{

library(readr)

GIVE EXAMPLES

}
}
