% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_cloglog_output.R
\name{rescale_cloglog_suit}
\alias{rescale_cloglog_suit}
\title{Rescales cloglog suitability output from MaxEnt model as an exponential.}
\usage{
rescale_cloglog_suit(xy.predicted, rescale.name, thresh, summary.file)
}
\arguments{
\item{xy.predicted}{Data import. The predicted cloglog suitability output
taken from one of the internal functions:
\code{\link[=predict_xy_suitability]{predict_xy_suitability()}} or
\code{\link[=predict_xy_suitability_CV]{predict_xy_suitability_CV()}}. Should be a data.frame or .csv
file. If an import, file path should be in the format produced by the
\code{\link[=file.path]{file.path()}} function (i.e. with '/' instead of '\\').}

\item{rescale.name}{Character. Descriptive name to be given to column of
rescaled values and the output object name. Exclude unnecessary phrases.}

\item{thresh}{thresh Numeric or Character. This may be imported manually
(numeric), or may be selected from one of the thresholds for the model
(character). See details for a list of preset options and other usages.}

\item{summary.file}{Data import. Contains preset values for thresh. Should be
a .csv file or data frame that contains the summary statistics output created by
\code{\link[=compute_MaxEnt_summary_statistics]{compute_MaxEnt_summary_statistics()}} or
\code{\link[=compute_MaxEnt_summary_statistics_CV]{compute_MaxEnt_summary_statistics_CV()}} The filename should end in
"summary_all_iterations.csv"). If an import, file path should be in the format
produced by the \code{\link[=file.path]{file.path()}} function (i.e. with '/' instead of '\\').}
}
\value{
Returns the input data frame, with the column of cloglog suitability values
replaced by the re-scaled exponential values.
}
\description{
Rescales cloglog suitability output from MaxEnt model as an exponential.
}
\details{
Requires the following packages: 'tidyverse', 'cli'.
\subsection{thresh:}{

Thresh presets list:
\itemize{
\item \code{BTO} = Balance training omission predicted area and threshold value
\item \code{EE} = Equate entropy of thresholded and original distributions
\item \code{ETSS} = Equal training sensitivity and specificity
\item \code{MTP} = Minimum Training Presence
\item \code{MTSS} = Maximum training sensitivity plus specificity
\item \code{ten_percentile} or \verb{10_percentile} = Ten percentile training presence
}
}
}
\examples{

example

}
