% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_model_output.R
\name{weight_model_output}
\alias{weight_model_output}
\title{Weight model output}
\usage{
weight_model_output(x)
}
\arguments{
\item{x}{Data import. Should be applied on across the cells of a raster, so x
can be thought of as the imported value of a single cell.}
}
\value{
returns weight
}
\description{
Weight model output
}
\details{
details
}
\examples{

example

}
