% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_MaxEnt_fixed_ROC.R
\name{compute_MaxEnt_fixed_ROC}
\alias{compute_MaxEnt_fixed_ROC}
\title{Create a model output}
\usage{
compute_MaxEnt_fixed_ROC(
  model.obj,
  model.name,
  mypath,
  fixed.area.points,
  env.covar.obj,
  test.obj
)
}
\arguments{
\item{model.obj}{}

\item{model.name}{Character. A string matching the name of the object set for
\code{model.obj}. Exclude unnecessary phrases, such as the "_model" ending.}

\item{mypath}{Character.A file path to the sub directory where the model
output will be stored. Should be used with the \code{\link[=file.path]{file.path()}} function
(i.e. with '/' instead of '\\').}

\item{fixed.area.points}{Data import. Should be a .csv file or data frame
that contains the fixed area points. If a .csv file, file path should be in the format
produced by the \code{\link[=file.path]{file.path()}} function (i.e. with '/' instead of '\\').
See details for additional import formatting.}

\item{env.covar.obj}{}

\item{test.obj}{}
}
\value{
blarg
}
\description{
The function requires the packages 'tidyverse', 'here', 'devtools', 'viridis', 'pROC', 'SDMtune', and 'patchwork'
}
\details{
\subsection{fixed.area.points:}{

The fixed area points dataset should be be in the following format: 3 columns,
with x as longitude, y as latitude and cell as the cell number from the
raster, in that order. Original dataset was generated using
\code{dismo::randomPoints()}
}
}
\examples{

TBD

}
