% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_MaxEnt_summary_statistics.R
\name{compute_MaxEnt_summary_statistics}
\alias{compute_MaxEnt_summary_statistics}
\title{Create a model output}
\usage{
compute_MaxEnt_summary_statistics(
  model.obj,
  model.name,
  mypath,
  create.dir = FALSE,
  env.covar.obj,
  train.obj,
  trainFolds.obj,
  test.obj,
  plot.types,
  threshold.types
)
}
\arguments{
\item{model.obj}{}

\item{model.name}{Character. A string matching the name of the object set for
\code{model.obj}. Exclude unnecessary phrases, such as the "_model" ending.}

\item{mypath}{Character.A file path to the sub directory where the model
output will be stored. Should be used with the \code{\link[=file.path]{file.path()}} function
(i.e. with '/' instead of '\\'). If this sub directory does not already exist
and should be created by the function, set \code{create.dir} = TRUE. This will
create a folder from the last part of the filepath in mypath.}

\item{create.dir}{Logical. Should the last element of \code{mypath} create a sub
directory for the model output? If TRUE, the main folder will be created for
the model output. If FALSE (ie, the sub directory already exists), only the
"plots" folder within the model output sub directory will be created.}

\item{env.covar.obj}{}

\item{train.obj}{}

\item{trainFolds.obj}{}

\item{test.obj}{}

\item{plot.types}{}

\item{threshold.types}{}
}
\value{
All MaxEnt Summary Statistics
}
\description{
The function requires the packages 'tidyverse', 'here', 'devtools', 'rJava', 'dismo', 'SDMtune', 'viridis', 'plotROC' and 'terra'.
}
\details{
model.obj <- entire_easternUSA_model
model.name <- "entire_easternUSA_model"
mypath <- file.path(here() \%>\%
dirname(),
"maxent/models/slf_easternUSA_entire_step1")
env.covar.obj <- x_env_covariates
train.obj <- entire_easternUSA_train
trainFolds.obj <- entire_easternUSA_trainFolds
test.obj <- entire_easternUSA_test
plot.types <- c("train", "test") # used to produce jackknife plots
threshold.types <- c("cloglog", "logistic") # used to produce marginal and univariate response curves
}
\examples{

TBD

}
