% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_risk_report.R
\name{create_risk_report}
\alias{create_risk_report}
\title{Creates}
\usage{
create_risk_report(
  locality,
  locality.type,
  save.report = FALSE,
  mypath,
  create.dir = FALSE,
  map.style = NA
)
}
\arguments{
\item{locality}{The name of the country, state or province for which to
generate the report. Please type the full name and not an abbreviation
(ex: United States, not USA). Avoid special characters,
but please include those used used in the ethnic naming (ex: Côte d'Ivoire).}

\item{locality.type}{Character. One of "country" or "states_provinces". If you do
not know the state or province you are looking for, you might create a report
at the country level and then look at the return for the name of the provinces
/ states included.}

\item{save.report}{Logical. Should the report be saved to file? File location
specified by \code{mypath}.}

\item{mypath}{Character.A file path to the sub directory where the model
output will be stored. Should be used with the \code{\link[=file.path]{file.path()}} function
(i.e. with '/' instead of '\\'). If this sub directory does not already exist
and should be created by the function, set \code{create.dir} = TRUE. This will
create a folder from the last part of the filepath in \code{mypath}.}

\item{create.dir}{Logical. Should the last element of \code{mypath} create a sub
directory for the report? If TRUE, the main folder will be created for
the model output. If FALSE (ie, the sub directory already exists), no directory
will be created.}

\item{map.style}{List, default is NA. This is used to apply ggplot aesthetics
to the mapped outputs. If specified, the given value should be a list of
ggplot aesthetic options. If not, the built-in default list will be used
(see details). See examples for usage.}
}
\value{
returns a report in list format. If save.report = TRUE, will also save the
report to file at destination specified by \code{mypath}. The outputs in the returned list
object include:
\itemize{
\item viticultural_regions_list - a list of known important wine regions within the locality
\item risk_maps - a current and future map of risk for L delicatula establishment based on 4 climate models
\item viticultural_risk_plot = a quantified assessment of the risk for L delicatula establishment for known wine regions within the locality.
\item viticultural_risk_table = a table quantifying the risk plot for viticultural areas
\item range_shift_map = a map of potential range expansion for L delicatula under climate change
}

Some maps may be formatted strangely because of a country's outlying territories.
You may need to further crop the plot using xlim and ylim (see examples).
}
\description{
blarg
}
\details{
Requires the following packages: 'tidyverse', 'terra', 'sf', 'here', 'cli', 'rnaturalearth', 'rnaturalearthhires', 'kableExtra', 'formattable', 'webshot2'.

Note that this function performs downloads from \href{https://www.naturalearthdata.com/}{naturalearthdata.com}.
The function will automatically create subfolders in \code{root/data-raw}
containing the shapefiles.

This function depends on certain files that have been distributed with this
package, which will be imported from \code{root/R/sysdata.rda} when the function
is run. The code to create \code{sysdata.rda} can be found in
\verb{root/vignettes/010_initialize_pkg.R}.

Here is a list of the files included in \code{create_risk_report_import.RData}:
\itemize{
\item global_model_summary.rds                                            | created in vignette 050
\item ensemble_thresh_values.rds                                          | created in vignette 110
\item wineries_tidied.rds                                                 | created in vignette 130
\item regional_ensemble_wineries_1981-2010_xy_pred_suit.rds               | created in vignette 130
\item regional_ensemble_wineries_2041-2070_GFDL_ssp370_xy_pred_suit.rds   | created in vignette 130
\item global_wineries_1981-2010_xy_pred_suit.rds                          | created in vignette 130
\item global_wineries_2041-2070_GFDL_ssp370_xy_pred_suit.rds              | created in vignette 130
}

Additionally, 3 rasters are used to create maps. These rasters are located in
\code{root/vignette-outputs/rasters}, These are the files:
\itemize{
\item slf_binarized_summed_1981-2010.asc                                  | created in vignette 120
\item slf_binarized_summed_2041-2070_ssp370_GFDL.asc                      | created in vignette 120
\item slf_range_shift_summed.asc                                          | created in vignette 120
}
\subsection{locality}{

\emph{Note} If you want results for Alaska or Hawaii, please use the settings
\code{locality = 'akaska'} and \code{locality.type = 'states_provinces'}.
}
}
\examples{

The output is in list format, so it should be called using this notation:

# find viticultural regions in locality
viticultural_regions <- slf_risk_report[[2]]

# alternatively, call elements by name:
risk_map_2055 <- slf_risk_report[["risk_maps"]][["2055_risk_map"]]


}
