% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_xy_suitability.R
\name{predict_xy_suitability}
\alias{predict_xy_suitability}
\title{Predicts suitability for xy coordinates according to a trained MaxEnt model (SDMmodelCV object)}
\usage{
predict_xy_suitability(
  xy.obj,
  xy.type,
  env.covar.obj,
  model.obj,
  mypath,
  predict.fun = "mean",
  clamp.pred = TRUE,
  output.name
)
}
\arguments{
\item{xy.obj}{Should be a .csv file or data frame that contains xy coordinate
data only. Coordinates should be in decimal degree format. Longitude should be
listed first, followed by latitude. If a .csv file, file path should be in the
format produced by the \code{\link[=file.path]{file.path()}} function (i.e. with '/' instead of '\\').}

\item{xy.type}{Character. Description of the xy data. If species occurrence
data, list the name of the species. If another type of coordinate data, list that
name. If providing more than one data type or species, separate into different
data frames before adding to this function}

\item{env.covar.obj}{A stack of rasters of environmental covariates. These
covariates may be the same covariates used to train the model, or they may be
temporally or spatially projected (ex, for climate change). Suitability
predictions will be made based on these rasters. See details for additional
formatting information.}

\item{model.obj}{A model object created by the package \code{SDMtune}.}

\item{mypath}{Character.A file path to the sub directory where the model
output will be stored. Should be used with the \code{\link[=file.path]{file.path()}} function
(i.e. with '/' instead of '\\'). If this sub directory does not already exist
and should be created by the function, set \code{create.dir = TRUE}. This will
create a folder from the last part of the filepath in \code{mypath}.}

\item{predict.fun}{Character. The default is \code{mean}. This is the function to
be applied to combine the iterations of the model when predicting a suitability
output. Can be one of: \code{min}, \code{mean}, \code{median}, \code{max}, or \code{sd}
(standard deviation). If multiple are desired, must be in the concatenated
form: \code{c("mean", "max")}. Should be all lowercase.}

\item{output.name}{The name of the file output. Separate words with _. Relevant
information might include the name of the model used to predict, the spatial
scale, the temporal scale, and the type of data points that are being used for
predictions.}
}
\value{
Returns 3 .csv files. The first file \code{xy_with_data} contains the input xy
coordinates with the value of each given environmental covariate at that
location. The second file \code{xy_no_data} contains the contains any
coordinates that could not be predicted because of an NA value for one or more
covariates at that location. The result file \code{xy_predicted_suitability}
contains each coordinate, the type of record and the predicted cloglog
suitability value.
}
\description{
This function predicts cloglog establishment suitability based on a trained
MaxEnt model for a set of xy coordinates. These coordinates do not need to be
within the training area for the model.
}
\details{
The function requires the packages 'cli', 'tidyverse', and 'SDMtune'.
}
\examples{
examples


}
