% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_cloglog_suitability.R
\name{rescale_cloglog_suitability}
\alias{rescale_cloglog_suitability}
\title{Rescales cloglog suitability output from MaxEnt model as an exponential.}
\usage{
rescale_cloglog_suitability(
  xy.predicted,
  thresh,
  summary.file,
  rescale.name = NA,
  rescale.thresholds = FALSE
)
}
\arguments{
\item{xy.predicted}{Data import. The predicted cloglog suitability output
taken from one of the internal package functions:
\code{\link[=predict_xy_suitability]{predict_xy_suitability()}} or
\code{\link[=predict_xy_suitability_CV]{predict_xy_suitability_CV()}}. Should be a data.frame or .csv
file. If an import, file path should be in the format produced by the
\code{\link[=file.path]{file.path()}} function (i.e. with '/' instead of '\\').}

\item{thresh}{Numeric or Character. This may be imported manually
(numeric), or may be selected from one of the thresholds for the model
(character). See details for a list of preset options and other usages.}

\item{summary.file}{Data import. Contains preset values for thresh and summary
statistics and is created by the internal package funcions:
\code{\link[=compute_MaxEnt_summary_statistics]{compute_MaxEnt_summary_statistics()}} or
\code{\link[=compute_MaxEnt_summary_statistics_CV]{compute_MaxEnt_summary_statistics_CV()}} The filename should end in
"summary_all_iterations.csv"). Should be a .csv file or data frame. If an
import, file path should be in the format produced by the \code{\link[=file.path]{file.path()}}
function (i.e. with '/' instead of '\\').}

\item{rescale.name}{Character. Descriptive name to be given to column of
rescaled values.}

\item{rescale.thresholds}{Logical. If true, the function will also rescale
the list of thesholds given in the thresh presets list. All other thresholds
are rescaled with reference to the threshold specified by \code{thresh}. This is
advised if visualizing with multiple thresholds.}
}
\value{
Returns the input data frame, with the column of cloglog suitability values
replaced by the re-scaled exponential values.

If rescale.thresholds = TRUE, will return a list of data frames. The first
object in the list will be the rescaled suitability values, and the second
object in the list will be the rescaled thresholds.
}
\description{
This function will take the cloglog suitability output from MaxEnt, which is
on a 0-1 scale, and re-scale it to have a set median value. The range of the
values will still be 0-1, but the median will now be the value of the \code{thresh}
parameter. The scaling function is applied as an exponential, for the purposes
of visualizing suitability change around a critical threshold (at the scale of
0-1, it can often be hard to see changes across the suitability threshold if
its value is very small). Note that this function will not work if the value
of thresh is 0.
}
\details{
This function was co-authored by Jacob Woods.

Requires the following packages: 'tidyverse', 'cli'.
\subsection{thresh:}{

Thresh presets list:
\itemize{
\item \code{BTO} = Balance training omission predicted area and threshold value
\item \code{EE} = Equate entropy of thresholded and original distributions
\item \code{ETSS} = Equal training sensitivity and specificity
\item \code{MTP} = Minimum Training Presence
\item \code{MTSS} = Maximum training sensitivity plus specificity
\item \code{ten_percentile} or \verb{10_percentile} = Ten percentile training presence
}
}
}
\examples{

example

}
